% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{abs.value}
\alias{abs.value}
\title{Absolute value for value objects}
\usage{
\method{abs}{value}(x)
}
\arguments{
\item{x}{A value object}
}
\value{
A new value object representing abs(x)
}
\description{
Element-wise absolute value.
}
