% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{backward.default}
\alias{backward.default}
\title{Default implementation does not propagate gradients. For instance, if we
have a constant, then the partial of the constant is not meaningful.}
\usage{
\method{backward}{default}(x, ...)
}
\arguments{
\item{...}{pass additional arguments}

\item{e}{A value object for which the backward pass should be performed}
}
\description{
Default implementation does not propagate gradients. For instance, if we
have a constant, then the partial of the constant is not meaningful.
}
