% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{backward}
\alias{backward}
\title{Generic function for the Backward pass for automatic differentiation (finds
the gradient of every sub-node in the computational graph with respect to
\code{e}). In other words, it is responsible for computing the gradient with
respect to \code{e}.}
\usage{
backward(e, ...)
}
\arguments{
\item{e}{An object for which the backward pass should be performed}

\item{...}{additional arguments to pass}
}
\description{
This generic function should be implemented for specific classes. We provide
an implementation for \code{value} objects.
}
