% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{backward.value}
\alias{backward.value}
\title{Backward pass for value objects}
\usage{
\method{backward}{value}(x)
}
\arguments{
\item{e}{A value object for which the backward pass should be performed}
}
\description{
Performs the backward pass (gradient computation) for a value object in the
computational graph with automatic differentiation.
}
