% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{contains}
\alias{contains}
\title{Set membership test for value objects.}
\usage{
contains(xs, obj)
}
\arguments{
\item{xs}{A list of value objects}

\item{obj}{A value object}
}
\value{
Boolean, determine if \code{xs} has \code{obj}
}
\description{
Set membership test for value objects.
}
\keyword{internal}
