% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{cos.value}
\alias{cos.value}
\title{Cosine function for value objects}
\usage{
\method{cos}{value}(x)
}
\arguments{
\item{x}{A value object}
}
\value{
A new value object representing cos(x)
}
\description{
Element-wise cosine.
}
