% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{data<-}
\alias{data<-}
\title{Set the data of a value object}
\usage{
data(x) <- value
}
\arguments{
\item{x}{A value object}

\item{value}{The new numeric value to assign (converted to matrix)}
}
\value{
The modified value object (invisibly)
}
\description{
Assignment function to update the data field of a value object
without breaking the computational graph reference.
}
\examples{
\dontrun{
x <- val(5)
data(x) <- 10
data(x)  # Returns 1x1 matrix containing 10
}

}
