% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{data.default}
\alias{data.default}
\title{Default implementation for retrieving the data from a differentiable object}
\usage{
\method{data}{default}(x, ...)
}
\arguments{
\item{x}{A non-value object}

\item{...}{pass additional arguments}
}
\value{
The original object
}
\description{
Default implementation for retrieving the data from a differentiable object
}
