% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{data}
\alias{data}
\title{Retrieve the data stored by an object.}
\usage{
data(x, ...)
}
\arguments{
\item{x}{An object to retrieve the value or data from}

\item{...}{additional arguments to pass}
}
\value{
The value or data of a differnetiable object
}
\description{
Retrieve the data stored by an object.
}
