% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{data.value}
\alias{data.value}
\title{Retrieve the value or data from a value object}
\usage{
\method{data}{value}(x, ...)
}
\arguments{
\item{x}{A value object}

\item{...}{additional arguments to pass}
}
\value{
The value or data of the value object
}
\description{
Retrieve the value or data from a value object
}
