% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{digamma.value}
\alias{digamma.value}
\title{Digamma (psi) function for value objects}
\usage{
\method{digamma}{value}(x)
}
\arguments{
\item{x}{A value object}
}
\value{
A new value object representing digamma(x)
}
\description{
Element-wise digamma: d/dx log(gamma(x)).
}
