% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{dof}
\alias{dof}
\alias{dof.hypothesis_test}
\title{Extract degrees of freedom from hypothesis test}
\usage{
dof(x, ...)

\method{dof}{hypothesis_test}(x, ...)
}
\arguments{
\item{x}{A hypothesis test object}

\item{...}{Additional arguments (ignored)}
}
\value{
Degrees of freedom
}
\description{
Extract degrees of freedom from hypothesis test
}
\section{Methods (by class)}{
\itemize{
\item \code{dof(hypothesis_test)}: Degrees of freedom for hypothesis tests

}}
