% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{-.value}
\alias{-.value}
\title{Subtraction for value objects}
\usage{
\method{-}{value}(x, y)
}
\arguments{
\item{x}{A value object (minuend)}

\item{y}{A value object or a scalar (subtrahend)}
}
\value{
A new value object representing the subtraction x - y
}
\description{
Subtraction for value objects
}
