% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{dual_num}
\alias{dual_num}
\title{Create a dual number}
\usage{
dual_num(primal, tangent = 0)
}
\arguments{
\item{primal}{The primal value}

\item{tangent}{The tangent (directional derivative), default 0}
}
\value{
A dual object
}
\description{
Create a dual number
}
