% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{exp.value}
\alias{exp.value}
\title{Exponential function for value objects}
\usage{
\method{exp}{value}(x)
}
\arguments{
\item{x}{A value object}
}
\value{
A new value object representing the exponential of x
}
\description{
Exponential function for value objects
}
