% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{fisher_information}
\alias{fisher_information}
\title{Compute observed Fisher information matrix}
\usage{
fisher_information(loglik_fn, params)
}
\arguments{
\item{loglik_fn}{Log-likelihood function taking a list of value parameters}

\item{params}{List of value objects at MLE}
}
\value{
The observed Fisher information matrix (negative Hessian)
}
\description{
For a log-likelihood function, the observed Fisher information is
the negative Hessian evaluated at the MLE. This is used for computing
standard errors of MLEs.
}
\details{
The observed Fisher information I(θ̂) = -H(θ̂) where H is the Hessian
of the log-likelihood. Standard errors are sqrt(diag(I⁻¹)).
}
