% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{grad}
\alias{grad}
\title{Gradient of \code{x} with respect to \code{e} in \code{backward(e)},
e.g., dx/de. (applies the chain rule)}
\usage{
grad(x, ...)
}
\arguments{
\item{x}{A differential object}

\item{...}{pass additional arguments}
}
\value{
The gradient as previously defined
}
\description{
Gradient of \code{x} with respect to \code{e} in \code{backward(e)},
e.g., dx/de. (applies the chain rule)
}
