% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{gradient}
\alias{gradient}
\title{Compute gradient as a numeric vector}
\usage{
gradient(loss_fn, params)
}
\arguments{
\item{loss_fn}{A function taking a list of value parameters}

\item{params}{A list of value objects}
}
\value{
A numeric vector of gradients
}
\description{
Convenience function to compute the gradient of a loss function
at the current parameter values.
}
