% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{inv_bounded}
\alias{inv_bounded}
\title{Inverse of bounded transform}
\usage{
inv_bounded(x, lower, upper)
}
\arguments{
\item{x}{Value in (lower, upper)}

\item{lower}{Lower bound}

\item{upper}{Upper bound}
}
\value{
Unconstrained value
}
\description{
Converts a bounded value back to unconstrained scale.
}
