% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{inv_positive}
\alias{inv_positive}
\title{Inverse of positive transform}
\usage{
inv_positive(x)
}
\arguments{
\item{x}{Positive value}
}
\value{
Unconstrained value (log of x)
}
\description{
Converts a positive value back to unconstrained scale.
}
