% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{inv_probability}
\alias{inv_probability}
\title{Inverse of probability transform}
\usage{
inv_probability(x)
}
\arguments{
\item{x}{Value in (0, 1)}
}
\value{
Unconstrained value (logit of x)
}
\description{
Converts a probability to unconstrained scale (logit).
}
