% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{is_dual}
\alias{is_dual}
\title{Check if object is a dual number}
\usage{
is_dual(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
TRUE if x is a dual object
}
\description{
Check if object is a dual number
}
