% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{is_value}
\alias{is_value}
\title{Check if an object is of class value}
\usage{
is_value(x)
}
\arguments{
\item{x}{The object to be checked}
}
\value{
TRUE if the object is of class value, FALSE otherwise
}
\description{
Determines if the input object is an instance of the value R6 class.
}
