% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{lgamma.value}
\alias{lgamma.value}
\title{Log-gamma function for value objects}
\usage{
\method{lgamma}{value}(x)
}
\arguments{
\item{x}{A value object}
}
\value{
A new value object representing lgamma(x)
}
\description{
Element-wise log(gamma(x)).
}
