% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{log.value}
\alias{log.value}
\title{Natural logarithm for value objects}
\usage{
\method{log}{value}(x, ...)
}
\arguments{
\item{x}{A value object}
}
\value{
A new value object representing log(x)
}
\description{
Element-wise natural logarithm.
}
