% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{logit}
\alias{logit}
\title{Logit function for value objects}
\usage{
logit(x)
}
\arguments{
\item{x}{A value object representing probabilities}
}
\value{
A new value object representing logit(x)
}
\description{
Element-wise logit: log(p/(1-p)).
}
