% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{loglik_bernoulli}
\alias{loglik_bernoulli}
\title{Bernoulli distribution log-likelihood}
\usage{
loglik_bernoulli(p, x)
}
\arguments{
\item{p}{Success probability (value object), must be in (0,1)}

\item{x}{Binary vector (0 or 1)}
}
\value{
A value object representing the log-likelihood
}
\description{
Special case of binomial with size=1.
L(p|x) = Σ\link{xᵢ*log(p) + (1-xᵢ)*log(1-p)}
}
