% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{loglik_beta}
\alias{loglik_beta}
\title{Beta distribution log-likelihood}
\usage{
loglik_beta(alpha, beta, x)
}
\arguments{
\item{alpha}{Shape parameter α (value object), must be positive}

\item{beta}{Shape parameter β (value object), must be positive}

\item{x}{Numeric vector of observations in (0,1)}
}
\value{
A value object representing the log-likelihood
}
\description{
Computes the log-likelihood for i.i.d. beta observations.
L(α,β|x) = n*\link{log(Γ(α+β)) - log(Γ(α)) - log(Γ(β))} +
(α-1)*Σlog(xᵢ) + (β-1)*Σlog(1-xᵢ)
}
