% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{loglik_binomial}
\alias{loglik_binomial}
\title{Binomial distribution log-likelihood}
\usage{
loglik_binomial(p, x, size)
}
\arguments{
\item{p}{Success probability (value object), must be in (0,1)}

\item{x}{Integer vector of successes}

\item{size}{Integer vector of trial counts (or single value if constant)}
}
\value{
A value object representing the log-likelihood
}
\description{
Computes the log-likelihood for binomial observations.
L(p|x,n) = Σ\link{xᵢ*log(p) + (nᵢ-xᵢ)*log(1-p) + log(C(nᵢ,xᵢ))}
}
\examples{
\dontrun{
x <- rbinom(50, size = 10, prob = 0.3)
p <- val(0.5)
ll <- loglik_binomial(p, x, size = 10)
backward(ll)
# MLE is sum(x) / sum(size)
}

}
