% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{loglik_exponential}
\alias{loglik_exponential}
\title{Exponential distribution log-likelihood}
\usage{
loglik_exponential(rate, x)
}
\arguments{
\item{rate}{Rate parameter λ (value object), must be positive}

\item{x}{Numeric vector of observations (must be non-negative)}
}
\value{
A value object representing the log-likelihood
}
\description{
Computes the log-likelihood for i.i.d. exponential observations.
L(λ|x) = n\emph{log(λ) - λ}Σxᵢ
}
\examples{
\dontrun{
x <- rexp(50, rate = 2)
rate <- val(1)
ll <- loglik_exponential(rate, x)
backward(ll)
grad(rate)  # should be n/rate - sum(x)
}

}
