% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{loglik_normal}
\alias{loglik_normal}
\title{Normal (Gaussian) log-likelihood}
\usage{
loglik_normal(mu, sigma, x)
}
\arguments{
\item{mu}{Mean parameter (value object)}

\item{sigma}{Standard deviation parameter (value object), must be positive}

\item{x}{Numeric vector of observations}
}
\value{
A value object representing the log-likelihood
}
\description{
Computes the log-likelihood for i.i.d. normal observations.
L(μ,σ²|x) = -n/2 log(2π) - n/2 log(σ²) - Σ(xᵢ-μ)²/(2σ²)
}
\examples{
\dontrun{
x <- rnorm(100, mean = 5, sd = 2)
mu <- val(0)
sigma <- val(1)
ll <- loglik_normal(mu, sigma, x)
backward(ll)
grad(mu)  # score for mu
}

}
