% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{loglik_poisson}
\alias{loglik_poisson}
\title{Poisson distribution log-likelihood}
\usage{
loglik_poisson(lambda, x)
}
\arguments{
\item{lambda}{Rate parameter λ (value object), must be positive}

\item{x}{Integer vector of observations (counts)}
}
\value{
A value object representing the log-likelihood
}
\description{
Computes the log-likelihood for i.i.d. Poisson observations.
L(λ|x) = Σxᵢ\emph{log(λ) - n}λ - Σlog(xᵢ!)
}
\details{
The term Σlog(xᵢ!) is constant w.r.t. λ and is included for completeness.
}
\examples{
\dontrun{
x <- rpois(100, lambda = 3)
lambda <- val(1)
ll <- loglik_poisson(lambda, x)
backward(ll)
# MLE is mean(x)
}

}
