% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{observed_info}
\alias{observed_info}
\alias{observed_info.femtofit}
\title{Observed Fisher information matrix}
\usage{
observed_info(object, ...)

\method{observed_info}{femtofit}(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments.}
}
\value{
The observed information matrix.
}
\description{
Returns the observed Fisher information matrix, which is the
negative Hessian of the log-likelihood at the MLE.
}
\section{Methods (by class)}{
\itemize{
\item \code{observed_info(femtofit)}: Observed information for femtofit

}}
