% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{optimization}
\alias{optimization}
\title{Optimization routines for maximum likelihood estimation}
\description{
These functions find the optimum (maximum or minimum) of differentiable
objective functions using femtograd's automatic differentiation.
}
