% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{^.value}
\alias{^.value}
\title{Power operation for value objects.}
\usage{
\method{^}{value}(b, e)
}
\arguments{
\item{b}{A value object as the base}

\item{e}{A value object, a scalar, or a numeric constant as the exponent}
}
\value{
A new value object representing the power b^e
}
\description{
Power operation for value objects.
}
