% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{primal}
\alias{primal}
\title{Extract primal from dual or return value unchanged}
\usage{
primal(x)
}
\arguments{
\item{x}{A dual or non-dual object}
}
\value{
The primal value
}
\description{
Extract primal from dual or return value unchanged
}
