% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{print.bootstrap_result}
\alias{print.bootstrap_result}
\title{Print method for bootstrap results}
\usage{
\method{print}{bootstrap_result}(x, ...)
}
\arguments{
\item{x}{A \code{bootstrap_result} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for bootstrap results
}
