% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.hessian_check}
\alias{print.hessian_check}
\alias{print.convergence_check}
\alias{print.model_diagnostics}
\title{Print methods for diagnostic objects}
\usage{
\method{print}{hessian_check}(x, ...)

\method{print}{convergence_check}(x, ...)

\method{print}{model_diagnostics}(x, ...)
}
\arguments{
\item{x}{A diagnostic object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print methods for diagnostic objects
}
