% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{print.likelihood_ratio_test}
\alias{print.likelihood_ratio_test}
\title{Print method for likelihood ratio test}
\usage{
\method{print}{likelihood_ratio_test}(x, ...)
}
\arguments{
\item{x}{A likelihood_ratio_test object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for likelihood ratio test
}
