% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{print.profile_likelihood}
\alias{print.profile_likelihood}
\title{Print method for profile likelihood}
\usage{
\method{print}{profile_likelihood}(x, ...)
}
\arguments{
\item{x}{A \code{profile_likelihood} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for profile likelihood
}
