% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{print.wald_test}
\alias{print.wald_test}
\title{Print method for Wald test}
\usage{
\method{print}{wald_test}(x, ...)
}
\arguments{
\item{x}{A wald_test object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for Wald test
}
