% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{pval}
\alias{pval}
\alias{pval.hypothesis_test}
\title{Extract p-value from hypothesis test}
\usage{
pval(x, ...)

\method{pval}{hypothesis_test}(x, ...)
}
\arguments{
\item{x}{A hypothesis test object}

\item{...}{Additional arguments (ignored)}
}
\value{
The p-value
}
\description{
Extract p-value from hypothesis test
}
\section{Methods (by class)}{
\itemize{
\item \code{pval(hypothesis_test)}: p-value for hypothesis tests

}}
