% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{relu}
\alias{relu}
\title{ReLU (Rectified Linear Unit) activation function for value objects}
\usage{
relu(x)
}
\arguments{
\item{x}{A value object}
}
\value{
A new value object representing max(0, x)
}
\description{
ReLU (Rectified Linear Unit) activation function for value objects
}
