% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{repr}
\alias{repr}
\title{Representation of a value object}
\usage{
repr(x)
}
\arguments{
\item{x}{A value object}
}
\value{
A string representation of the value object
}
\description{
Creates a string representation of a value object including its data and gradient values.
}
