% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{retrieve.default}
\alias{retrieve.default}
\title{Retrieve the value or data from a non-value object}
\usage{
\method{retrieve}{default}(x)
}
\arguments{
\item{x}{A non-value object}
}
\value{
The original object
}
\description{
Retrieve the value or data from a non-value object
}
