% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{retrieve}
\alias{retrieve}
\title{Retrieve the value or data from an object}
\usage{
retrieve(x)
}
\arguments{
\item{x}{An object to retrieve the value or data from}
}
\value{
The value or data of the object if it is a value object, or the original object if it is not a value object
}
\description{
Retrieve the value or data from an object
}
