% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{retrieve.value}
\alias{retrieve.value}
\title{Retrieve the value or data from a value object}
\usage{
\method{retrieve}{value}(x)
}
\arguments{
\item{x}{A value object}
}
\value{
The value or data of the value object
}
\description{
Retrieve the value or data from a value object
}
