% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{se}
\alias{se}
\alias{se.femtofit}
\title{Standard errors from a fitted model}
\usage{
se(object, ...)

\method{se}{femtofit}(object, ...)
}
\arguments{
\item{object}{A \code{femtofit} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Named numeric vector of standard errors.
}
\description{
Extracts standard errors of coefficient estimates.
}
\section{Methods (by class)}{
\itemize{
\item \code{se(femtofit)}: Standard errors for femtofit

}}
