% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid activation function for value objects}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{A value object}
}
\value{
A new value object representing sigmoid(x)
}
\description{
Element-wise sigmoid: 1/(1+exp(-x))
}
