% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{sin.value}
\alias{sin.value}
\title{Sine function for value objects}
\usage{
\method{sin}{value}(x)
}
\arguments{
\item{x}{A value object}
}
\value{
A new value object representing sin(x)
}
\description{
Element-wise sine.
}
