% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{/.value}
\alias{/.value}
\title{Division for value objects}
\usage{
\method{/}{value}(x, y)
}
\arguments{
\item{x}{A value object (numerator)}

\item{y}{A value object or a scalar (denominator)}
}
\value{
A new value object representing the division x / y
}
\description{
Division for value objects
}
