% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{sqrt.value}
\alias{sqrt.value}
\title{Square root for value objects}
\usage{
\method{sqrt}{value}(x)
}
\arguments{
\item{x}{A value object}
}
\value{
A new value object representing the square root of x
}
\description{
Square root for value objects
}
