% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{sum_values}
\alias{sum_values}
\title{Summation for a list of value objects}
\usage{
sum_values(values = NULL)
}
\arguments{
\item{values}{A list of objects that can be added}
}
\value{
A new value object representing the sum of all input value objects
}
\description{
Summation for a list of value objects
}
