% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{summary.bootstrap_result}
\alias{summary.bootstrap_result}
\title{Summary for bootstrap results}
\usage{
\method{summary}{bootstrap_result}(object, level = 0.95, ...)
}
\arguments{
\item{object}{A \code{bootstrap_result} object.}

\item{level}{Confidence level (default 0.95).}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary for bootstrap results
}
