% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.R
\name{tangent}
\alias{tangent}
\title{Extract tangent from dual or return 0}
\usage{
tangent(x)
}
\arguments{
\item{x}{A dual or non-dual object}
}
\value{
The tangent value
}
\description{
Extract tangent from dual or return 0
}
