% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis.R
\name{test_stat}
\alias{test_stat}
\alias{test_stat.hypothesis_test}
\title{Extract test statistic from hypothesis test}
\usage{
test_stat(x, ...)

\method{test_stat}{hypothesis_test}(x, ...)
}
\arguments{
\item{x}{A hypothesis test object}

\item{...}{Additional arguments (ignored)}
}
\value{
The test statistic
}
\description{
Extract test statistic from hypothesis test
}
\section{Methods (by class)}{
\itemize{
\item \code{test_stat(hypothesis_test)}: Test statistic for hypothesis tests

}}
