% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{topological_sort}
\alias{topological_sort}
\title{Build the topological order for a given value object}
\usage{
topological_sort(v)
}
\arguments{
\item{v}{A value object}
}
\value{
A list representing the topological order of nodes
}
\description{
Build the topological order for a given value object
}
\keyword{internal}
