% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{upper_bounded}
\alias{upper_bounded}
\title{Transform to upper-bounded interval}
\usage{
upper_bounded(x, upper)
}
\arguments{
\item{x}{Unconstrained value}

\item{upper}{Upper bound}
}
\value{
Value < upper
}
\description{
Maps unconstrained values to (-∞, upper).
Equivalent to upper - positive(x).
}
\details{
The transformation is: upper_bounded(x, b) = b - exp(x)
}
\examples{
\dontrun{
# Parameter < 10
raw <- val(0)
param <- upper_bounded(raw, 10)  # 10 - exp(0) = 9
}

}
