% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{val}
\alias{val}
\title{\code{value} object constructor}
\usage{
val(data, label = "")
}
\arguments{
\item{data}{Numeric value of the object}

\item{label}{Optional character label for debugging

Represents a value node in the computational graph with automatic
differentiation. This constructor function creates a new value object with
the specified data and optional label.}
}
\description{
\code{value} object constructor
}
