% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{vcov_matrix}
\alias{vcov_matrix}
\title{Compute variance-covariance matrix from Hessian}
\usage{
vcov_matrix(hess, is_loglik = TRUE)
}
\arguments{
\item{hess}{The Hessian matrix}

\item{is_loglik}{If TRUE, returns -H⁻¹ (for log-likelihood)}
}
\value{
The variance-covariance matrix
}
\description{
Compute variance-covariance matrix from Hessian
}
