% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{backward}
\alias{backward}
\title{Backward pass for automatic differentiation}
\usage{
backward(e)
}
\arguments{
\item{e}{An object for which the backward pass should be performed}
}
\description{
Generic function for the backward pass, responsible for computing gradients in the
computational graph. This function should be implemented for specific classes.
}
