% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{data.value}
\alias{data.value}
\title{Retrieve the data matrix from a value object}
\usage{
\method{data}{value}(x, drop = TRUE, ...)
}
\arguments{
\item{x}{A value object}

\item{drop}{If TRUE (default) and result is 1x1, return scalar.
Set to FALSE to always return a matrix.}

\item{...}{additional arguments to pass}
}
\value{
The data as scalar (if 1x1 and drop=TRUE) or matrix
}
\description{
Retrieve the data matrix from a value object
}
