% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{grad.default}
\alias{grad.default}
\title{Default gradient is zero matrix}
\usage{
\method{grad}{default}(x, ...)
}
\arguments{
\item{x}{A non-value object}

\item{...}{pass additional arguments}
}
\value{
Zero matrix of appropriate dimensions
}
\description{
Default gradient is zero matrix
}
