% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{grad.value}
\alias{grad.value}
\title{Gradient of a \code{value} object \code{x} with respect to \code{e} in
\code{backward(e)}, e.g., dx/de. (applies the chain rule)}
\usage{
\method{grad}{value}(x, ...)
}
\arguments{
\item{x}{A value object}

\item{...}{pass additional arguments}
}
\value{
The value or data of the value object
}
\description{
Gradient of a \code{value} object \code{x} with respect to \code{e} in
\code{backward(e)}, e.g., dx/de. (applies the chain rule)
}
