% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{log1p.value}
\alias{log1p.value}
\title{Log(1+x) for value objects}
\usage{
\method{log1p}{value}(x)
}
\arguments{
\item{x}{A value object}
}
\value{
A new value object representing log(1+x)
}
\description{
Element-wise log(1+x), numerically stable for small x.
}
