% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{relu}
\alias{relu}
\title{ReLU activation function for value objects}
\usage{
relu(x)
}
\arguments{
\item{x}{A value object}
}
\value{
A new value object representing max(0, x)
}
\description{
Element-wise ReLU: max(0, x)
}
