% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{sum.value}
\alias{sum.value}
\title{Summation for value objects}
\usage{
\method{sum}{value}(..., na.rm = FALSE)
}
\arguments{
\item{...}{value objects and/or numeric values to sum}

\item{na.rm}{Logical, whether to remove NA values}
}
\value{
A new value object (1x1 matrix) representing the sum
}
\description{
Sums all elements of value objects, returning a 1x1 value.
}
