% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{tanh.value}
\alias{tanh.value}
\title{Hyperbolic tangent for value objects}
\usage{
\method{tanh}{value}(x)
}
\arguments{
\item{x}{A value object}
}
\value{
A new value object representing tanh(x)
}
\description{
Element-wise tanh.
}
