% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{*.value}
\alias{*.value}
\title{Multiplication for value objects}
\usage{
\method{*}{value}(e1, e2)
}
\arguments{
\item{e1}{A value object}

\item{e2}{A value object or a scalar}
}
\value{
A new value object representing the product
}
\description{
Multiplication for value objects
}
