% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{val}
\alias{val}
\title{\code{value} object constructor}
\usage{
val(data)
}
\arguments{
\item{data}{Numeric value of the object

Represents a value node in the computational graph with automatic
differentiation. This constructor function creates a new value object with
the specified data.}
}
\description{
\code{value} object constructor
}
