% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{value}
\alias{value}
\title{value R6 class}
\description{
value R6 class

value R6 class
}
\details{
Represents a node in the computational graph with automatic differentiation.
Each value object holds data, gradient, a backward function, previous nodes,
and an operation name.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Numeric value of the object}

\item{\code{grad}}{Gradient value, initially set to 0}

\item{\code{backward_fn}}{A function that performs the backward pass (gradient computation)}

\item{\code{prev}}{A list of previous nodes in the computational graph}

\item{\code{op}}{A character string representing the operation associated with the node}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-value-new}{\code{value$new()}}
\item \href{#method-value-clone}{\code{value$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-value-new"></a>}}
\if{latex}{\out{\hypertarget{method-value-new}{}}}
\subsection{Method \code{new()}}{
Initializes a new value object with the given data, list of children, and operation name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{value$new(data, children = list(), op = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Numeric value of the object}

\item{\code{children}}{List of previous nodes in the computational graph (default: empty list)}

\item{\code{op}}{A character string representing the operation associated with the node (default: empty string)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-value-clone"></a>}}
\if{latex}{\out{\hypertarget{method-value-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{value$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
